Combine the PromoListingRepository with the PromoRepository since it's redudant and an extra class when PromoRepository suffices, update areas to comply with the new combination
