+++ contracts/Repositories/PromoRepositoryContract.php
@@ -10,4 +10,11 @@ interface PromoRepositoryContract
      * @return array
      */
     public function getHomepagePromos(int $page_id);
+
+    /**
+     * Get promotions for the listing.
+     *
+     * @return array
+     */
+    public function getPromoListingPromos(array $data, $limit = 75);
 }
